Revert "refresh spark new data"

+++ sql/catalyst/src/main/java/org/apache/spark/sql/catalyst/expressions/xml/UDFXPathUtil.java
@@ -180,7 +180,7 @@ public class UDFXPathUtil {
         return 0;
       }
       // Bound skip by beginning and end of the source
-      int n = (int) Math.min(length - next, ns);
+      long n = Math.min(length - next, ns);
       n = Math.max(-next, n);
       next += n;
       return n;
